First pass working OpenAL system

+++ Core/FCTypes.h
@@ -23,7 +23,7 @@
 #import "FCColor.h"
 #import "FCKeys.h"
 
-typedef uint64_t FCHandle;	// should last a while 8)
+typedef uint32_t FCHandle;	// should last a while 8)
 
 static const FCHandle kFCHandleInvalid = 0;
 static const FCHandle kFCHandleFirstValid = 1;
